+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Package Linux toolbox with runtime dependencies.
+- Create system independent toolbox with all supported system's files.
 
 ### Fixed